+++ src/io/mod.rs
@@ -1,0 +1,2 @@
+mod save;
+pub use save::SaveData;
