+++ src/model/mod.rs
@@ -1,0 +1,2 @@
+mod connection;
+pub use connection::{Connection, Servers};
